Change default url to app-agents.contrastsecurity.com/Contrast, move autofac tests
